% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANOVA.R
\name{ANOVA}
\alias{ANOVA}
\title{Title ANOVA}
\usage{
ANOVA(file, species_col, rain_col)
}
\arguments{
\item{file}{the file path}

\item{species_col}{provide column you wish to use}

\item{rain_col}{state the "Cumulative_Rain" to be used}
}
\value{
analysis of the butterfly data and plots the outcome
}
\description{
Title ANOVA
}
