% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeanWeight.R
\name{avg_weight}
\alias{avg_weight}
\title{Mean weight of wood rats based off sex and year}
\usage{
avg_weight(dataframe, weight, sex, year)
}
\arguments{
\item{dataframe}{The path to the csv file you want to process}

\item{weight}{The weight of wood rats in grams}

\item{sex}{The sex of wood rats, male or female}

\item{year}{The year the data was recorded}
}
\value{
Average weight of wood rats for each year of data that was taken based off sex
}
\description{
Mean weight of wood rats based off sex and year
}
