% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubble.R
\name{bubble}
\alias{bubble}
\title{Title bubbles}
\usage{
bubble(file, column)
}
\arguments{
\item{file}{give the function a file path}

\item{column}{give the function the column you wish to observe}
}
\value{
Problem values
}
\description{
Title bubbles
}
