% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{Title clean dataframe}
\usage{
clean(file_path)
}
\arguments{
\item{file_path}{gives the function a file_path to use}
}
\value{
a cleaned data set
}
\description{
Title clean dataframe
}
