% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter_Function.R
\name{Filter_Function}
\alias{Filter_Function}
\title{Filters out blanks, NAs, and juveniles of the data}
\usage{
Filter_Function(data)
}
\arguments{
\item{data}{The data you want to clean}
}
\value{
Returns a clean dataset
}
\description{
Filters out blanks, NAs, and juveniles of the data
}
