% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Frequency_function.R
\name{Frequency_function}
\alias{Frequency_function}
\title{Title Generates the total number of samples of multiple columns}
\usage{
Frequency_function(data, column1, column2, column3)
}
\arguments{
\item{data}{Dataframe your data is from}

\item{column1}{The first column of data you want to sum}

\item{column2}{The second column of data you want to sum}

\item{column3}{The third column of data you want to sum}
}
\value{
the total number of entries from your chosen columns
}
\description{
Title Generates the total number of samples of multiple columns
}
