% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_2.R
\name{Graph_Function}
\alias{Graph_Function}
\title{Graph Function}
\usage{
Graph_Function(Dataframe, Xvar, Yvar)
}
\arguments{
\item{Dataframe}{Data file you wish to use}

\item{Xvar}{X variable for the graph}

\item{Yvar}{Y variable for the graph}
}
\value{
A simple scatter plot
}
\description{
Graph Function
}
