% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteractionPlot.R
\name{interaction_plot}
\alias{interaction_plot}
\title{Interaction plot to show correlation between mean weight of wood rats for each year and sex}
\usage{
interaction_plot(dataframe, weight, sex, year)
}
\arguments{
\item{dataframe}{The path to the csv file you want to process}

\item{weight}{The weight of wood rats in grams}

\item{sex}{The sex of wood rats, male or female}

\item{year}{The year the data was recorded}
}
\value{
Interaction plot of mean weight of wood rats per year for males and females
}
\description{
Interaction plot to show correlation between mean weight of wood rats for each year and sex
}
