% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearModel.R
\name{linear_model}
\alias{linear_model}
\title{Linear Model of mean weight of wood rats based on sex and year}
\usage{
linear_model(dataframe, weight, sex, year)
}
\arguments{
\item{dataframe}{The path the csv file you want to process}

\item{weight}{The weight of wood rats in grams}

\item{sex}{The sex of wood rats, male or female}

\item{year}{The year the data was recorded}
}
\value{
Linear model of data for mean weight of wood rats based off year and sex
}
\description{
Linear Model of mean weight of wood rats based on sex and year
}
