% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linear_Plot.R
\name{Linear_plot_function}
\alias{Linear_plot_function}
\title{Linear Plot Function: creates a linear plot when fed two variables}
\usage{
Linear_plot_function(
  data,
  x = column1,
  y = column2,
  xlab = "text1",
  ylab = "text2",
  title = "text3",
  label = "text4"
)
}
\arguments{
\item{data}{The desired dataframe}

\item{x}{The predictor}

\item{y}{The response}

\item{xlab}{Desired text label for the x axis should be encased by quotes}

\item{ylab}{Desired text label for the y axis should be encased by quotes}

\item{title}{Desired text label for the chart title should be encased by quotes}

\item{label}{Desired text label for the R square value should be encased by quotes}
}
\value{
A linear regression plot with labels
}
\description{
Linear Plot Function: creates a linear plot when fed two variables
}
