% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LM_Test_Function.R
\name{LM_Function}
\alias{LM_Function}
\title{Crates various linear model tests to see if your data is normally disturbed}
\usage{
LM_Function(data, responce_var, predictor_var)
}
\arguments{
\item{data}{The dataset you wish to use}

\item{responce_var}{Variable you want to select as the response}

\item{predictor_var}{Variable you want to select as the predictor}
}
\value{
Creates a summery of the linear model and Shapiro-Wilk test
}
\description{
Crates various linear model tests to see if your data is normally disturbed
}
