% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LM_Graph_Function.R
\name{LM_Graph_Function}
\alias{LM_Graph_Function}
\title{Creates a linear model graph}
\usage{
LM_Graph_Function(data, x_var, y_var)
}
\arguments{
\item{data}{The dataset you wish to use for the graph}

\item{x_var}{The X/response variable you want to use for the graph}

\item{y_var}{The Y/predictor variable you want to use for the graph}
}
\value{
A graph showing the linear relationship between the chosen variables
}
\description{
Creates a linear model graph
}
