% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Map.R
\name{mapping_function}
\alias{mapping_function}
\title{Creates a map from an excel sheet containing three columns latitude,longitude,locality}
\usage{
mapping_function(data, column1, column2, column3)
}
\arguments{
\item{data}{The dataframe of interest}

\item{column1}{Longitude}

\item{column2}{Latitude}

\item{column3}{Locality}
}
\value{
A map from the Leaflet function
}
\description{
Creates a map from an excel sheet containing three columns latitude,longitude,locality
}
