% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanra.R
\name{meanr}
\alias{meanr}
\title{Calculate means of a given column in dataframe}
\usage{
meanr(path, column)
}
\arguments{
\item{path}{Path to a comma-separated values file}

\item{column}{The column for which you'd like the mean}
}
\value{
result A numerical value representing the mean
}
\description{
Calculate means of a given column in dataframe
}
