% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/My_Function.R
\name{my_function}
\alias{my_function}
\title{Checks to see if fish in selected column are too small}
\usage{
my_function(filepath, columns)
}
\arguments{
\item{filepath}{The path to the file you want to process}

\item{columns}{The selected column you want to use in the function}
}
\value{
column indicating fish that are too small. Will print "Check small value"
}
\description{
Checks to see if fish in selected column are too small
}
