% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanr.R
\name{new_function}
\alias{new_function}
\title{Remove NAs and prints the values that are greater than five in a column}
\usage{
new_function(filepath, column)
}
\arguments{
\item{filepath}{The path to the tab_delimited file you want to use}

\item{column}{what column you wish to filter}
}
\value{
The problem values
}
\description{
Remove NAs and prints the values that are greater than five in a column
}
