% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoxPlot.R
\name{plot_func}
\alias{plot_func}
\title{Box plot of wood rats weight for each year}
\usage{
plot_func(dataframe, weight, sex, year)
}
\arguments{
\item{dataframe}{The path to the csv file you want to process}

\item{weight}{The weight of wood rats in grams}

\item{sex}{The sex of wood rat, male or female}

\item{year}{The year the data was recorded}
}
\value{
Box plot of weight of wood rats per year for males and females
}
\description{
Box plot of wood rats weight for each year
}
