% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/practice_function.R
\name{practice_function}
\alias{practice_function}
\title{Takes in a csv database and squares the values of a numeric column specified}
\usage{
practice_function(filepath, clmns)
}
\arguments{
\item{filepath}{The path to the csv file you want to process}

\item{clmns}{The column of values you wish to get squared values for}
}
\value{
sqrd the original dataset with a new column of squared values.
}
\description{
Takes in a csv database and squares the values of a numeric column specified
}
