% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_trace.R
\name{process_trace}
\alias{process_trace}
\title{Remove burnin and plot posterior trace}
\usage{
process_trace(filepath, burnin = 0.25)
}
\arguments{
\item{filepath}{The path to the tab-delimited file you want to process}

\item{burnin}{The proportion of the trace to discard as burnin}
}
\value{
plot A ggplot object of the posterior trace
}
\description{
Remove burnin and plot posterior trace
}
