% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ratio_Function.R
\name{Ratio_Function}
\alias{Ratio_Function}
\title{Calculates three ratios of a selected columns and adds a new columns the dataset with those values}
\usage{
Ratio_Function(data, column1, column2, column3)
}
\arguments{
\item{data}{The dataset you want to use}

\item{column1}{The column that you want to select for the calculation in this case SL}

\item{column2}{The column that you want to select for the calculation in this case LL}

\item{column3}{The column that you want to select for the calculation in this case TL}
}
\value{
A new dataset with an added columns of the three calculated ratios
}
\description{
Calculates three ratios of a selected columns and adds a new columns the dataset with those values
}
