% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemoveNAs.R
\name{remove_nas}
\alias{remove_nas}
\title{Remove NAs from surveys dataset}
\usage{
remove_nas(dataframe, weight, sex, year)
}
\arguments{
\item{dataframe}{The path to the csv file you want to process}

\item{weight}{The weight of wood rats in grams}
}
\value{
surveys dataset with removed NAs and "Not Recorded" values
}
\description{
Remove NAs from surveys dataset
}
