% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasons.R
\name{season}
\alias{season}
\title{Title Season}
\usage{
season(file)
}
\arguments{
\item{df}{this is the data frame to be used.}
}
\value{
an overview of the butterfly data related to the Summer/Spring
}
\description{
Title Season
}
