% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_lm.R
\name{survey_lm}
\alias{survey_lm}
\title{Takes in the surveys dataset, cleans the data, and performs a linear model analysis and plot for specified columns.}
\usage{
survey_lm(filepath, predictor, response, x_title, y_title)
}
\arguments{
\item{filepath}{The path to the csv file you want to process}

\item{predictor}{The predictor variable from the csv filepath inputted}

\item{response}{The response variable from the csv filepath inputted}

\item{x_title}{User-specified title for x-axis of plot}

\item{y_title}{User-specified title for y-axis of plot}
}
\value{
survey_plot Returns a scatterplot of detailing the linear model predicting change in the response variable based on a change in the predictor
}
\description{
The surveys dataset is installed and cleaned utilizing the survey_cleaner function.
Then the values of the predictor variable are extracted to the variable predict.
Also, the values of the response variable are extracted to the variable respond.
A linear model is fit for the relationship of the response variable with changes to the predictor using the lm function saved to survey_model.
A scatterplot of the linear relationship would created and saved under surveys_lm_plot.png to check for proper data input.
The function lastly returns a summary of the linear model test to get the relevant values from the statistical analysis.
}
