% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_three_function_two.R
\name{surveys_plot}
\alias{surveys_plot}
\title{Graph of filtered data}
\usage{
surveys_plot(output)
}
\arguments{
\item{output}{name of function used to filter data}
}
\value{
plot1 plot of filtered data
}
\description{
Graph mean weight by year of species in surveys data
}
