% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_three_function_one.R
\name{surveys_weight_by_year}
\alias{surveys_weight_by_year}
\title{Filters data from surveys}
\usage{
surveys_weight_by_year(surveys)
}
\arguments{
\item{surveys}{a dataset containing a column you'd like to summarize}
}
\value{
weights a summary of the data set
}
\description{
Filters the mean weight by year from surveys data
}
