% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thesis_avg_function.R
\name{Thesis_avg_function}
\alias{Thesis_avg_function}
\title{Averages the data in a given column}
\usage{
Thesis_avg_function(data, columns)
}
\arguments{
\item{data}{The dataframe of interest}

\item{columns}{The column that you wish to average}
}
\value{
the average of the column in integer form
}
\description{
Averages the data in a given column
}
