% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubble.R
\name{bubble}
\alias{bubble}
\title{Title bubbles}
\usage{
bubble(filepath, column)
}
\arguments{
\item{filepath}{Give the function the filepath}

\item{column}{give the functions a column you want to observe}
}
\value{
Problem values
}
\description{
Title bubbles
}
