% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Violin_Plot_Function.R
\name{Graph_Function}
\alias{Graph_Function}
\title{Creates violin plot}
\usage{
Graph_Function(data, Xvar, Yvar, fill_data, facet_var, Xtext, Ytext)
}
\arguments{
\item{data}{The dataset you want to use when making this graph}

\item{Xvar}{X variable for the graph}

\item{Yvar}{Y variable for the graph}

\item{fill_data}{Variable that fills the points}

\item{facet_var}{Variable you can want to facet the graph by}

\item{Xtext}{Label for the X axis}

\item{Ytext}{Label for the Y axis}
}
\value{
A violin plot that is faceted by your specific grouping
}
\description{
Creates violin plot
}
