% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_function2.R
\name{new_function2}
\alias{new_function2}
\title{Multiplies hindfoot length and weight}
\usage{
new_function2(dataframe, hindfoot_length, weight)
}
\arguments{
\item{dataframe}{The path to the tab-delimited file you want to process}

\item{hindfoot_length}{The length of wood rats' hindfoot in mm}

\item{weight}{The weight of wood rats in grams}
}
\value{
Error message for numbers greater than 5500
}
\description{
Multiplies hindfoot length and weight
}
