% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squared_surveys.R
\name{squared_surveys}
\alias{squared_surveys}
\title{Takes in a csv database and squares the values of a numeric column specified}
\usage{
squared_surveys(filepath, clmns)
}
\arguments{
\item{filepath}{The path to the csv file you want to process}

\item{clmns}{The column of values you wish to get squared values for}
}
\value{
sqrd the squared values of the original column.
}
\description{
Takes in a csv database and squares the values of a numeric column specified
}
