% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_anova.R
\name{survey_anova}
\alias{survey_anova}
\title{Takes in the surveys dataset from a given filepath and performs an ANOVA analysis of the data for two given column types.}
\usage{
survey_anova(filepath, predictor, response)
}
\arguments{
\item{filepath}{The path to the csv file you want to process}

\item{predictor}{The predictor variable from the csv filepath inputted}

\item{response}{The response variable from the csv filepath inputted}
}
\value{
anova_plot Returns a scatterplot of the anova model predicting change in the response variable based on a change in the predictor
}
\description{
Takes in the surveys dataset from a given filepath and performs an ANOVA analysis of the data for two given column types.
}
