% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_cleaner.R
\name{survey_cleaner}
\alias{survey_cleaner}
\title{Takes in the surveys dataset filepath, checks if the file is not downloaded it downloads, and then rids the data of any type of non-applicable values returning the altered csv dataset.}
\usage{
survey_cleaner(filepath)
}
\arguments{
\item{filepath}{The path to the csv file you want to process}
}
\value{
surveys the cleaned dataset with no values labeled NA, "Not Recorded", or an empty space.
}
\description{
Takes in the surveys dataset filepath, checks if the file is not downloaded it downloads, and then rids the data of any type of non-applicable values returning the altered csv dataset.
}
