% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_lm.R
\name{survey_lm}
\alias{survey_lm}
\title{Takes in the surveys dataset, cleans the data, and performs a linear model analysis and plot for specified columns.}
\usage{
survey_lm(filepath, predictor, response)
}
\arguments{
\item{filepath}{The path to the csv file you want to process}

\item{predictor}{The predictor variable from the csv filepath inputted}

\item{response}{The response variable from the csv filepath inputted}
}
\value{
survey_plot Returns a scatterplot of detailing the linear model predicting change in the response variable based on a change in the predictor
}
\description{
Takes in the surveys dataset, cleans the data, and performs a linear model analysis and plot for specified columns.
}
